// src/app/shared/models/vehicle.model.ts
export interface Vehicle {
  vehicleID?: number;
  vehicleNumber: string;
  vehicleType: string;
  chassisNumber: string;
  engineNumber: string;
  manufacturingYear: number;
  loadCarryingCapacity: number;
  makeOfVehicle: string;
  modelNumber: string;
  bodyType: string;
  organisationName: string;
  deviceID: string;
}